Module uic_image_provider::image_watcher

Structs

IMAGE_MAP

ImageEntry

ImageList

ImageWatcher

MapEntry

MetaFile

TwoStringsKey

Enums

WatchError

Statics

IMG_PATH

Functions

calc_md5_digest
Calculate and return md5 base64 encoded digest from buffer

error_message
Print error messages

file_exists
Check if file exists and return buffer with content

get_image_file
Get file from Map Should be implemented as queue But it looks mosquitto itself already has queue for on-message handler